+++ src/test/java/gift/RepositoryTest/OptionEntityRepositoryTest.java
@@ -1,11 +1,10 @@
-package gift;
+package gift.RepositoryTest;
 
 import gift.Model.Entity.*;
 import gift.Repository.*;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.event.annotation.BeforeTestClass;
 
 import java.util.ArrayList;
 import java.util.List;
