+++ src/test/java/gift/RepositoryTest/ProductEntityRepositoryTest.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.RepositoryTest;
 
 import gift.Model.Entity.CategoryEntity;
 import gift.Model.Entity.ProductEntity;
